+++ internal/client/control_plane.go
@@ -2,6 +2,7 @@ package client
 
 import (
 	"github.com/hashicorp/go-memdb"
+	babel "github.com/umbracle/babel/sdk"
 	"github.com/umbracle/vesta/internal/server/proto"
 )
 
@@ -11,4 +12,7 @@ type ControlPlane interface {
 
 	// UpdateAlloc updates an allocation state
 	UpdateAlloc(alloc *proto.Allocation) error
+
+	// UpdateSyncStatus updates the sync status of an allocation
+	UpdateSyncStatus(alloc, task string, status *babel.SyncStatus) error
 }